refactor: circlelayer and pointlayer components
