+++ giraffe/src/components/geo/PointMapLayer.tsx
@@ -1,7 +1,6 @@
 // Libraries
 import React, {useEffect, useState} from 'react'
 import {FunctionComponent, useLayoutEffect} from 'react'
-import L from 'leaflet'
 import {Marker} from 'react-leaflet'
 import MarkerClusterGroup from 'react-leaflet-markercluster'
 import 'react-leaflet-markercluster/dist/styles.min.css'
@@ -9,19 +8,18 @@ import 'react-leaflet-markercluster/dist/styles.min.css'
 // Utils
 import {getColor} from './dimensionCalculations'
 import {SVGIcon} from './SVGIcon'
-import {defineToolTipEffect, formatValue} from './processing/toolTips'
+import {defineToolTipEffect} from './processing/toolTips'
 import {GeoTooltip} from './GeoTooltip'
+import {
+  formatPointLayerRowInfo,
+  createClusterCustomIcon,
+  MARKER_ICON_SIZE,
+} from '../../utils/geo'
 
 // Types
 import {GeoTable} from './processing/GeoTable'
-import {ClusterAggregation, GeoPointMapViewLayer} from '../../types/geo'
+import {GeoPointMapViewLayer} from '../../types/geo'
 import {Config} from '../../types'
-
-// Constants
-const MARKER_ICON_SIZE = [26, 38]
-const CLUSTER_ICON_SIZE = 30
-const CLUSTER_ICON_SIZE_BORDER = 4
-const DEFAULT_CLUSTER_COLOR = '#00a3ff'
 
 interface Props {
   table: GeoTable
@@ -29,114 +27,6 @@ interface Props {
   properties: GeoPointMapViewLayer
   stylingConfig: Partial<Config>
   isClustered: boolean
-}
-
-const formatRowInfo = (
-  properties: GeoPointMapViewLayer,
-  table: GeoTable,
-  index
-) => {
-  const result = []
-  const timeValue = table.getTimeString(index)
-  if (timeValue) {
-    result.push({
-      key: '_time',
-      name: 'Time',
-      type: 'string',
-      values: [timeValue],
-    })
-  }
-  const colorValue = table.getValue(index, properties.colorField)
-  const {colorDimension} = properties
-  const colorInfo = formatValue(
-    properties.colorField,
-    'Color',
-    colorValue,
-    colorDimension
-  )
-  if (colorInfo) {
-    result.push(colorInfo)
-  }
-  return result
-}
-
-const getClusterValues = (cluster, result: number[]): GeoPointMapViewLayer => {
-  let clusterRenderingProperties
-  for (const marker of cluster._markers) {
-    result.push(marker.value)
-    clusterRenderingProperties = marker.clusterRenderingProperties
-  }
-  for (const childCluster of cluster._childClusters) {
-    const properties = getClusterValues(childCluster, result)
-    if (properties) {
-      clusterRenderingProperties = properties
-    }
-  }
-  return clusterRenderingProperties
-}
-
-const getClusterValue = (
-  cluster
-): {value: number; properties: GeoPointMapViewLayer} => {
-  const values: number[] = []
-  const properties = getClusterValues(cluster, values)
-  const aggregation =
-    properties.clusterAggregationFunction || ClusterAggregation.max
-  switch (aggregation) {
-    case ClusterAggregation.max:
-      return {value: Math.max(...values), properties}
-    case ClusterAggregation.min:
-      return {value: Math.min(...values), properties}
-    case ClusterAggregation.mean:
-      return {
-        value: values.reduce((a, x) => a + x, 0) / values.length,
-        properties,
-      }
-    case ClusterAggregation.median:
-      if (values.length === 1) {
-        return {value: values[0], properties}
-      }
-      values.sort()
-      if (values.length % 2 === 0) {
-        const half = values.length / 2
-        return {value: (values[half - 1] + values[half]) / 2, properties}
-      } else {
-        return {value: values[Math.floor(values.length / 2)], properties}
-      }
-    default:
-      return {value: 0, properties}
-  }
-}
-
-const createClusterCustomIcon = function(cluster) {
-  let outerColor = DEFAULT_CLUSTER_COLOR,
-    innerColor = 'white'
-  const {value, properties} = getClusterValue(cluster)
-  if (properties.areClustersColored && properties.colors) {
-    innerColor = 'white'
-    outerColor = getColor(properties.colors, value, false)
-  }
-  const outer = `\
-background-color: ${outerColor};\
-height: ${CLUSTER_ICON_SIZE}px;\
-border-radius: ${CLUSTER_ICON_SIZE / 2}px;`
-  const inner = `\
-height: ${CLUSTER_ICON_SIZE - 2 * CLUSTER_ICON_SIZE_BORDER}px;\
-width: ${CLUSTER_ICON_SIZE - 2 * CLUSTER_ICON_SIZE_BORDER}px;\
-left: ${CLUSTER_ICON_SIZE_BORDER}px;\
-top: ${CLUSTER_ICON_SIZE_BORDER}px;\
-position: relative;\
-border-radius:${(CLUSTER_ICON_SIZE - 2 * CLUSTER_ICON_SIZE_BORDER) / 2}px;\
-background-color: ${innerColor};\
-font-size: 13px;\
-text-align: center;\
-color: dimgray;\
-line-height:${CLUSTER_ICON_SIZE - 2 * CLUSTER_ICON_SIZE_BORDER}px;`
-  return L.divIcon({
-    html: `<div style="${outer}"><div style="${inner}">${cluster.getChildCount()}</div></div>`,
-    className: 'marker-cluster-custom',
-    iconSize: L.point(CLUSTER_ICON_SIZE, CLUSTER_ICON_SIZE, true),
-  })
 }
 
 export const PointMapLayer: FunctionComponent<Props> = props => {
@@ -169,7 +59,7 @@ export const PointMapLayer: FunctionComponent<Props> = props => {
         icon={icon}
       />
     )
-    const rowInfo = formatRowInfo(properties, table, i)
+    const rowInfo = formatPointLayerRowInfo(properties, table, i)
     tooltips.push({markerRef, rowInfo})
   }
   const tooltip = (
